fix: caffeine bloom, method set bit count

+++ shared/include/userver/cache/impl/frequency_sketch.hpp
@@ -288,7 +288,7 @@ FrequencySketch<T, Hash, FrequencySketchPolicy::CaffeineBloom>::SetBitCount(
     int64_t x) {
   x = x - ((x >> 1) & 0x5555555555555555);
   x = (x & 0x3333333333333333) + ((x >> 2) & 0x3333333333333333);
-  return (((x + (x >> 4)) & 0xF0F0F0F0F0F0F0F) * 0x101010101010101) >> 56;
+  return static_cast<uint32_t>((((x + (x >> 4)) & 0xF0F0F0F0F0F0F0F) * 0x101010101010101) >> 56);
 }
 
 template <typename T, typename Hash>
